Enum tagua_parser::internal::Error

pub enum Error<P, E = u32> {
    Code(ErrorKind<E>),
    Node(ErrorKind<E>, Box<Err<P, E>>),
    Position(ErrorKind<E>, P),
    NodePosition(ErrorKind<E>, P, Box<Err<P, E>>),
}

Contains the error that a parser can return

It can represent a linked list of errors, indicating the path taken in the parsing tree, with corresponding position in the input data. It depends on P, the input position (for a &[u8] parser, it would be a &[u8]), and E, the custom error type (by default, u32)

Variants
Code(ErrorKind<E>)

An error code, represented by an ErrorKind, which can contain a custom error code represented by E

Node(ErrorKind<E>, Box<Err<P, E>>)

An error code, and the next error

Position(ErrorKind<E>, P)

An error code, and the input position

NodePosition(ErrorKind<E>, P, Box<Err<P, E>>)

An error code, the input position and the next error