Module tagua_parser::rules::expressions::primaries

Group of primary expression rules.

The list of all primary expressions is provided by the PHP Language Specification in the Grammar chapter, Expressions section.

Enums

IntrinsicError
Intrinsic errors.

Functions

anonymous_function
Recognize an anonymous function.

array
Recognize an array.

intrinsic
Recognize all kind of intrinsics.

intrinsic_echo
Recognize an echo.

intrinsic_empty
Recognize an empty.

intrinsic_eval
Recognize an lazy evaluation.

intrinsic_exit
Recognize an exit.

intrinsic_isset
Recognize an exit.

intrinsic_list
Recognize a list.

intrinsic_print
Recognize a print.

intrinsic_unset
Recognize an unset.

primary
Recognize all kind of primary expressions.